Split login route in separate file
